feat: migrate controller.js & fix some types
